introduce modelNameFromPayloadKey and deprecate typeForRoot
Introduced in this commit is payloadKeyFromModelName, which allows you
to specify the outgoing root key for a model name in the JSON payload.

For example:

```javascript
RESTSerializer.extend({modelNameFromPayloadKey: dasherize});
```

```json
{
  "tacoParty": { }
}
```

becomes

```json
{
  "taco-party"
}
```

For symmetry, modelNameFromPayloadKey has been introduced to
look up the ember data record in the store given a JSON root
key. This used to be called pathForType, which has been deprecated.